ES2AntiPatterns/src/Animal.java
public class Animal {

    String nome;
    String classe;
    int idade;

    public Animal(String nome, String classe, int idade) {
        this.nome = nome;
        this.classe = classe;
        this.idade = idade;
    }

    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    public String getClasse() {
        return classe;
    }

    public void setClasse(String classe) {
        this.classe = classe;
    }

    public int getIdade() {
        return idade;
    }

    public void setIdade(int idade) {
        this.idade = idade;
    }
}
